backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-nextjs==2.4.0
djangorestframework==3.14.0
fonttools==4.49.0
frozenlist==1.4.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
websockets==12.0
yarl==1.9.4